Correct adding data-spy attribute to nav
Bootstrap Genesis previously negelected to remove the .no-js class on
the html element.

Therefore, the fallback CSS for .no-js was being applied, which
prevented this issue surfacing.  The data-spy attribute was not being
applied therefore the JavaScript version of the fixed nav was not
occurring.

This fixes the problem by correctly adding the data-spy attribute by

- correcting the filter hook
- correcting the string modification

See #10